Move pkg/client Client to pkg/client/metaclient.
* Adjust naming, and add interface to make it more consistent with
kubernetes client sets.
